[plot]: move some functions in iros25_plotting into one utils.py file.
